django-bike-backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djongo==1.2.31
pymongo==4.15.2
django-cors-headers==4.9.0
python-decouple==3.8
